src/requirement.txt
streamlit
Pillow
numpy
opencv-python
pandas
glob
